Enable Support for QT6
In the CMakeLists.txt add a branch to check of qt6 is available.
If it is, use that, otherwise assume QT5.

In qfsignalproxy slightly adapt call to constructor of QVariant.
QVariant(type, void*) signature has changed between QT5 and QT6.

in qfstore adapt call to constructor of QQmlListProperty.
Signature has changed between QT5.13 and QT6